requirements.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
charset-normalizer==3.4.4
distro==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.11.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.0.5
langchain-core==1.0.4
langchain-openai==1.0.2
langgraph==1.0.2
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.2
langgraph-sdk==0.2.9
langsmith==0.4.41
numpy==2.3.4
openai==2.7.1
orjson==3.11.4
ormsgpack==1.12.0
packaging==25.0
pydantic==2.12.4
pydantic_core==2.41.5
python-dotenv==1.2.1
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
sniffio==1.3.1
tenacity==9.1.2
tiktoken==0.12.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
xxhash==3.6.0
zstandard==0.25.0
